Add Minkowski distance metric with adjustable `p`
This is a 2D generalisation of the rules for vector space distances.

- When `p = 1` it is equivalant to Manhattan distance
- When `p = 2` it is equivalent to Euclidian distance
- When `p = ∞` it is equivalent to Chebyshev distance

Due to the limit point requiring `max` of either dimension, in practice
Chebyshev can’t be computed using this formula (same with the `min` when
exponents are negative).

Potential use cases for this are in creating interesting shapes based on
distance from a center point. Tweak `p` for grid effects like magic spell
animations, procedural generation, etc. Good values to experiment with
range between 0.25 and 4. See the Wikipedia article for diagrams of what
these shapes look like on a 2D grid.